ibft: add support to use iscsi_ibft module and log into all portals found in firmware
This adds the ibft sysfs module (iscsi_ibft) parsing support. It
was original done by Konard, but I have ported it to use the
sysfs.c helpers (add new ones and ported iscsi_sysfs.c too).

This patch also modifies iscsistart and iscsiadm to print/log
into all the portals found in firmware.

It also changes the behavior of

iscsiadm -m discovery -t fw

so that we create db records for what is found. It is not
fully hooked into the iface code, so it will use the
different inititor name, but it will not create a iface for
it (this means that if you have a record for the same portal that
is bound to a iface with a iname you can run into problems).

+++ usr/iscsi_sysfs.h
@@ -93,7 +93,7 @@ extern void iscsi_sysfs_set_device_online(int hostno, int target, int lun);
 extern void iscsi_sysfs_rescan_device(int hostno, int target, int lun);
 extern int iscsi_sysfs_for_each_device(int host_no, uint32_t sid,
 				void (* fn)(int host_no, int target, int lun));
-extern struct iscsi_transport *iscsi_sysfs_get_transport_by_hba(long host_no);
+extern struct iscsi_transport *iscsi_sysfs_get_transport_by_hba(uint32_t host_no);
 extern struct iscsi_transport *iscsi_sysfs_get_transport_by_session(char *sys_session);
 extern struct iscsi_transport *iscsi_sysfs_get_transport_by_sid(uint32_t sid);
 extern struct iscsi_transport *iscsi_sysfs_get_transport_by_name(char *transport_name);
